+++ src/typing_ex/handler/core.clj
@@ -4,6 +4,7 @@
    [ataraxy.response :as response]
    [buddy.hashers :as hashers]
    [clojure.string :as str]
+   [environ.core :refer [env]]
    [hato.client :as hc]
    [integrant.core :as ig]
    [ring.util.anti-forgery :refer [anti-forgery-field]]
@@ -12,7 +13,7 @@
    [typing-ex.boundary.drills  :as drills]
    [typing-ex.boundary.results :as results]
    [typing-ex.boundary.status  :as status]
-   [typing-ex.boundary.users   :as users]
+   #_[typing-ex.boundary.users   :as users]
    [typing-ex.view.page :as view]
    ))
 
@@ -42,17 +43,18 @@
     ;;(timbre/debug "find-user body" body)
     body))
 
-;;
-(defn auth? [db login password]
-  (let [;;ret (users/find-user-by-login db login)
-        ret (find-user login)]
-    (timbre/debug "auth?" login)
-    (and (some? ret)
-         (hashers/check password (:password ret)))))
+;; FIXME: env 以外、system をみてスイッチしたい
+(defn auth? [login password]
+  (if (env :tp-dev)
+    true
+    (let [ret (find-user login)]
+      ;; (timbre/debug "auth?" login)
+      (and (some? ret)
+           (hashers/check password (:password ret))))))
 
 (defmethod ig/init-key :typing-ex.handler.core/login-post [_ {:keys [db]}]
   (fn [{[_ {:strs [login password]}] :ataraxy/result}]
-    (if (and (seq login) (auth? db login password))
+    (if (and (seq login) (auth? login password))
       (do
         (timbre/debug "login success" login)
         (-> (redirect "/sum/1")